Extract QwordFusion from explain Summary
The 64-bit reassembly (endian word offsets, per-path folding, or-branch
fusion) moves to its own class; Summary composes it.
